Bug fixes
1: Character falling off or climbing ledge N will be moved to the location and orientation of ledge N-1

Problem: HipToLedgeDistance had a Max Value of 70.0, which triggered the ledge-grabbing behavior, even when we haven't traced a wall or ledge.

Fix: Reduce HipToLedgeDistance's Max Value to 0.0

2: On Ramp_StaticMesh2's thin side, Mannequin cannot grab its ledge.

Problem: In cases where 2 or more objects overlap the Sphere Tracer, when one object exits the overlap state canTrace will be set to false, which is not the behavior we want.

Fix: add a data structure to house references to the group of objects. On Object Overlap, add it to the structure, On Exit, remove the object. I chose a Set. When it's size is 0, canTrace is set to false. otherwise, canTrace is set to True. This fixed everything!